backend/requirements.txt
Django==5.0.2
psycopg2-binary==2.9.9
djangorestframework==3.14.0
kinde-python-sdk==1.2.3
python-dotenv==1.0.1
openai==1.12.0
django-cors-headers==4.5.0
django_cloudinary_storage==0.3.0
Pillow==11.0.0
uvicorn[standard]==0.30.6
channels==4.0.0
